+++ src/harness/mod.rs
@@ -1,0 +1,40 @@
+use crate::TestKey;
+
+mod build;
+mod check;
+mod generate;
+mod read;
+mod run;
+
+pub use build::{build_test, link_test};
+pub use check::check_test;
+pub use generate::generate_test_src;
+pub use read::read_tests;
+pub use run::{run_dynamic_test, WriteBuffer};
+
+/// The name of a test for pretty-printing.
+pub fn full_test_name(
+    TestKey {
+        test_name,
+        convention,
+        caller_id,
+        callee_id,
+        ..
+    }: &TestKey,
+) -> String {
+    format!("{test_name}::{convention}::{caller_id}_calls_{callee_id}")
+}
+
+/// The name of a subtest for pretty-printing.
+pub fn full_subtest_name(
+    TestKey {
+        test_name,
+        convention,
+        caller_id,
+        callee_id,
+        ..
+    }: &TestKey,
+    func_name: &str,
+) -> String {
+    format!("{test_name}::{convention}::{caller_id}_calls_{callee_id}::{func_name}")
+}
